backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
pymupdf
pydantic
